Added ElasticSearch client and calling making query against local index.

+++ src/Sfa.Eds.Das.Web/Models/SearchResults.cs
@@ -4,7 +4,7 @@ namespace Sfa.Eds.Das.Web.Models
 {
     public sealed class SearchResults
     {
-        public int TotalResults { get; set; }
+        public long TotalResults { get; set; }
 
         public IEnumerable<SearchResultsItem> Results { get; set; }
     }
